+++ clang-tools-extra/docs/ReleaseNotes.rst
@@ -67,6 +67,12 @@ The improvements are...
 Improvements to clang-tidy
 --------------------------
 
+New module
+^^^^^^^^^^
+- New module `llvmlibc`.
+
+  This module contains checks related to the LLVM-libc coding standards.
+
 New checks
 ^^^^^^^^^^
 
@@ -94,6 +100,12 @@ New checks
 
   Flags use of the `C` standard library functions ``memset``, ``memcpy`` and
   ``memcmp`` and similar derivatives on non-trivial types.
+
+- New :doc:`llvmlibc-restrict-system-libc-headers
+  <clang-tidy/checks/llvmlibc-restrict-system-libc-headers>` check.
+
+  Finds includes of system libc headers not provided by the compiler within
+  llvm-libc implementations.
 
 - New :doc:`objc-dealloc-in-category
   <clang-tidy/checks/objc-dealloc-in-category>` check.